+++ CHANGELOG.md
@@ -1,1 +1,17 @@
 #  Change Log
+
+## v1.0.34 - 18-July-2021
+**Changes** 
+### Fixed
+- bug fixes
+
+## v1.0.33 - 17-July-2021
+**Changes** 
+### Fixed
+- unable to add/load comments 
+
+
+## v1.0.32 - 16-July-2021
+**Changes** 
+### Fixed
+- add feature product under edit profile